Refactor SongInfoScreen component
- Split the original SongInfoScreen component into smaller, reusable components.
- Created AlbumArtFlip component to handle album art flip animation and display.
- Created AdditionalInfo component to manage additional audio feature information.
- Integrated AdditionalInfo component within AlbumArtFlip component.
- Updated SongInfoScreen to use the new AlbumArtFlip, TrackInfo, TrackDetails, and AudioFeatures components.

This refactoring improves the modularity and readability of the code, making it easier to maintain and extend.
